refactor: move jwt related function to another file, make the register return token (user will logged in after register)
