+++ backend/handlers/jwt.go
@@ -1,0 +1,135 @@
+package handlers
+
+import (
+	"crypto/rand"
+	"encoding/base64"
+	"ku-work/backend/model"
+	"log"
+	"net/http"
+	"os"
+	"time"
+
+	"github.com/gin-gonic/gin"
+	"github.com/golang-jwt/jwt/v5"
+	"gorm.io/gorm"
+)
+
+type JWTHandlers struct {
+	DB *gorm.DB
+	JWTSecret []byte
+}
+
+func NewJWTHandlers(db *gorm.DB) *JWTHandlers {
+	jwtSecret := []byte(os.Getenv("JWT_SECRET"))
+
+	if len(jwtSecret) == 0 {
+		log.Fatal("JWT_SECRET environment variable is not set")
+	}
+	return &JWTHandlers{
+		DB: db,
+		JWTSecret: jwtSecret,
+	}
+}
+
+// Generate JWT and Refresh Tokens
+func (h *JWTHandlers) generateTokens(userID uint) (string, string, error) {
+	// JWT Token
+	jwtClaims := &model.UserClaims{
+		UserID: userID,
+		RegisteredClaims: jwt.RegisteredClaims{
+			ExpiresAt: jwt.NewNumericDate(time.Now().Add(time.Minute * 15)), // JWT expires in 15 minutes.
+			IssuedAt:  jwt.NewNumericDate(time.Now()),
+		},
+	}
+	jwtToken := jwt.NewWithClaims(jwt.SigningMethodHS256, jwtClaims)
+	signedJwtToken, err := jwtToken.SignedString(h.JWTSecret)
+	if err != nil {
+		return "", "", err
+	}
+
+	// Refresh Token
+	refreshTokenBytes := make([]byte, 32)
+	rand.Read(refreshTokenBytes)
+	refreshTokenString := base64.URLEncoding.EncodeToString(refreshTokenBytes)
+	
+	// Create or update the refresh token in the database.
+	refreshTokenDB := model.RefreshToken{
+		UserID:    userID,
+		Token:     refreshTokenString,
+		ExpiresAt: time.Now().Add(time.Hour * 24 * 30), // Refresh token expires in 30 days.
+	}
+	
+	h.DB.Save(&refreshTokenDB)
+
+	return signedJwtToken, refreshTokenString, nil
+}
+
+// handle refresh token request
+func (h *JWTHandlers) RefreshTokenHandler(ctx *gin.Context) {
+	refreshToken, err := ctx.Cookie("refresh_token")
+	if err != nil {
+		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "Missing refresh token cookie"})
+		return
+	}
+	
+	var refreshTokenDB model.RefreshToken
+	if err := h.DB.Where("token = ?", refreshToken).First(&refreshTokenDB).Error; err != nil {
+		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid refresh token"})
+		return
+	}
+	
+	if refreshTokenDB.ExpiresAt.Before(time.Now()) {
+		h.DB.Delete(&refreshTokenDB)
+		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "Refresh token expired"})
+		return
+	}
+	
+	jwtToken, newRefreshToken, err := h.generateTokens(refreshTokenDB.UserID)
+	if err != nil {
+		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to generate new tokens"})
+		return
+	}
+	
+	h.DB.Delete(&refreshTokenDB)
+	
+	ctx.SetCookie("refresh_token", newRefreshToken, int(time.Hour * 24 * 30 / time.Second), "/", "", true, true)
+
+	ctx.JSON(http.StatusOK, gin.H{"token": jwtToken})
+}
+
+// LogoutHandler invalidates the user's session.
+func (h *JWTHandlers) LogoutHandler(ctx *gin.Context) {
+	refreshToken, err := ctx.Cookie("refresh_token")
+	if err != nil {
+		ctx.JSON(http.StatusOK, gin.H{"message": "Logged out successfully"})
+		return
+	}
+
+	var refreshTokenDB model.RefreshToken
+	if err := h.DB.Where("token = ?", refreshToken).First(&refreshTokenDB).Error; err == nil {
+		h.DB.Delete(&refreshTokenDB)
+	}
+
+	ctx.SetCookie("refresh_token", "", -1, "/", "", true, true)
+
+	ctx.JSON(http.StatusOK, gin.H{"message": "Logged out successfully"})
+}
+
+// Handle Token sending
+func (h *JWTHandlers) HandleToken(req_ctx *gin.Context, user model.User) *gin.Context {
+	ctx := req_ctx
+	jwtToken, refreshToken, err := h.generateTokens(user.ID)
+	if err != nil {
+		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to generate tokens"})
+		return ctx
+	}
+	
+	// Set the refresh token cookie with a max age of 30 days
+	maxAge := int(time.Hour * 24 * 30 / time.Second)
+	ctx.SetCookie("refresh_token", refreshToken, maxAge, "/", "", true, true)
+
+	// Return the JWT token in the JSON response body.
+	ctx.JSON(http.StatusOK, gin.H{"token": jwtToken})
+	
+	return ctx
+}
