+++ backend/handlers/routes.go
@@ -9,16 +9,17 @@ import (
 
 func SetupRoutes(router *gin.Engine, db *gorm.DB) {
 	// Initialize handlers
-	authHandlers := NewAuthHandlers(db)
+	localAuthHandlers := NewLocalAuthHandlers(db)
+	jwtHandler := NewJWTHandlers(db)
 	
-	
-	router.POST("/register", authHandlers.RegisterHandler)
-	router.POST("/login", authHandlers.LoginHandler)	
-	router.POST("/refresh", authHandlers.RefreshTokenHandler)
-	router.POST("/logout", authHandlers.LogoutHandler)
+	router.POST("/register", localAuthHandlers.RegisterHandler)
+	router.POST("/login", localAuthHandlers.LoginHandler)	
+	router.POST("/refresh", jwtHandler.RefreshTokenHandler)
+	router.POST("/logout", jwtHandler.LogoutHandler)
 
 	
-	protected := router.Use(middlewares.AuthMiddleware(authHandlers.JWTSecret))
+	// Authentication Protected Routes
+	protected := router.Use(middlewares.AuthMiddleware(jwtHandler.JWTSecret))
 	protected.GET("/protected", func(ctx *gin.Context) {
 		ctx.JSON(200, gin.H{"message": "Protected route"})
 	})
